blog/requirements.txt
appdirs==1.4.0
coverage==4.3.4
defusedxml==0.4.1
Django==1.10.5
django-activeurl==0.1.9
django-allauth==0.30.0
django-appconf==1.0.2
django-classy-tags==0.8.0
django-crispy-forms==1.6.1
django-markdown-deux==1.0.5
django-pagedown==0.1.1
django-rest-auth==0.9.0
djangorestframework==3.5.3
Faker==0.7.3
lxml==3.7.2
markdown2==2.3.2
mixer==5.6.6
oauthlib==2.0.1
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.6.2
py==1.4.32
pyparsing==2.1.10
pytest==3.0.6
pytest-cov==2.4.0
pytest-django==3.1.2
python-dateutil==2.6.0
python3-openid==3.0.10
requests==2.13.0
requests-oauthlib==0.7.0
six==1.10.0
